requirements/base.txt
Pillow==5.4.1
PyYAML==3.13

celery==4.1.1

django-activity-stream==0.6.5
django-celery==3.2.2
django-colorful==1.3
django-cors-headers==2.4.0
django-downloadview==1.9
django-formtools==2.1
django-pure-pagination==0.3.0
django-mathfilters==0.4.0
django-model-utils==3.1.2
django-mptt==0.9.1
django-qsstats-magic==1.0.0
django-solo==1.1.3
django-stronghold==0.3.0
django-widget-tweaks==1.4.3
djangorestframework==3.9.2
djangorestframework-recursive==0.1.2
drf-yasg==1.12.1

flanker==0.9.0
flex==6.13.2
furl==2.0.0
fusepy==3.0.1

gevent==1.3.7
graphviz==0.10.1
gunicorn==19.9.0

Jinja2==2.10

mock==2.0.0

node-semver==0.5.1

pathlib2==2.3.3
pycountry==18.5.26
PyPDF2==1.26.0
pyocr==0.6
python-dateutil==2.7.5
python-gnupg==0.3.9
python-magic==0.4.15
pytz==2018.7

requests==2.20.1

sh==1.12.14
swagger-spec-validator==2.4.1

whitenoise==4.1.2
